Fix for PytestCollectionWarning: cannot collect test class because it has a __init__ constructor
