+++ tests/end_to_end/test_suites/task_runner_tests.py
@@ -4,7 +4,7 @@
 import pytest
 import logging
 
-from tests.end_to_end.utils.common_fixtures import (
+from tests.end_to_end.utils.tr_common_fixtures import (
     fx_federation_tr,
     fx_federation_tr_dws,
 )
