+++ tests/end_to_end/utils/tr_common_fixtures.py
@@ -5,7 +5,6 @@ import pytest
 import collections
 import concurrent.futures
 import logging
-import numpy as np
 
 import tests.end_to_end.utils.constants as constants
 import tests.end_to_end.utils.federation_helper as fh
@@ -21,10 +20,6 @@ federation_fixture = collections.namedtuple(
     "model_owner, aggregator, collaborators, workspace_path, local_bind_path",
 )
 
-workflow_local_fixture = collections.namedtuple(
-    "workflow_local_fixture",
-    "aggregator, collaborators, runtime",
-)
 
 @pytest.fixture(scope="function")
 def fx_federation_tr(request):
@@ -234,137 +229,4 @@ def fx_federation_tr_dws(request):
         collaborators=collaborators,
         workspace_path=workspace_path,
         local_bind_path=local_bind_path,
-    )
-
-
-@pytest.fixture(scope="function")
-def fx_local_federated_workflow(request):
-    """
-    Fixture to set up a local federated workflow for testing.
-    This fixture initializes an `Aggregator` and sets up a list of collaborators
-    based on the number specified in the test configuration. It also configures
-    a `LocalRuntime` with the aggregator, collaborators, and an optional backend
-    if specified in the test configuration.
-    Args:
-        request (FixtureRequest): The pytest request object that provides access
-                                to the test configuration.
-    Yields:
-        LocalRuntime: An instance of `LocalRuntime` configured with the aggregator,
-                    collaborators, and backend.
-    """
-    # Import is done inline because Task Runner does not support importing below openfl packages
-    from openfl.experimental.workflow.interface import Aggregator, Collaborator
-    from openfl.experimental.workflow.runtime import LocalRuntime
-    from tests.end_to_end.utils.wf_helper import (
-        init_collaborator_private_attr_index,
-        init_collaborator_private_attr_name,
-        init_collaborate_pvt_attr_np,
-        init_agg_pvt_attr_np
-    )
-    collab_callback_func = request.param[0] if hasattr(request, 'param') and request.param else None
-    collab_value = request.param[1] if hasattr(request, 'param') and request.param else None
-    agg_callback_func = request.param[2] if hasattr(request, 'param') and request.param else None
-
-    # Get the callback functions from the locals using string
-    collab_callback_func_name = locals()[collab_callback_func] if collab_callback_func else None
-    agg_callback_func_name = locals()[agg_callback_func] if agg_callback_func else None
-    collaborators_list = []
-
-    if agg_callback_func_name:
-        aggregator = Aggregator( name="agg",
-                                private_attributes_callable=agg_callback_func_name)
-    else:
-        aggregator = Aggregator()
-
-    # Setup collaborators
-    for i in range(request.config.num_collaborators):
-        func_var = i if collab_value == "int" else f"collaborator{i}" if collab_value == "str" else None
-        collaborators_list.append(
-            Collaborator(
-                name=f"collaborator{i}",
-                private_attributes_callable=collab_callback_func_name,
-                param = func_var
-            )
-        )
-
-    backend = request.config.backend if hasattr(request.config, 'backend') else None
-    if backend:
-        local_runtime = LocalRuntime(aggregator=aggregator, collaborators=collaborators_list, backend=backend)
-    local_runtime = LocalRuntime(aggregator=aggregator, collaborators=collaborators_list)
-
-    # Return the federation fixture
-    return workflow_local_fixture(
-        aggregator=aggregator,
-        collaborators=collaborators_list,
-        runtime=local_runtime,
-    )
-
-
-@pytest.fixture(scope="function")
-def fx_local_federated_workflow_prvt_attr(request):
-    """
-    Fixture to set up a local federated workflow for testing.
-    This fixture initializes an `Aggregator` and sets up a list of collaborators
-    based on the number specified in the test configuration. It also configures
-    a `LocalRuntime` with the aggregator, collaborators, and an optional backend
-    if specified in the test configuration.
-    Args:
-        request (FixtureRequest): The pytest request object that provides access
-                                to the test configuration.
-    Yields:
-        LocalRuntime: An instance of `LocalRuntime` configured with the aggregator,
-                    collaborators, and backend.
-    """
-    # Import is done inline because Task Runner does not support importing below openfl packages
-    from openfl.experimental.workflow.interface import Aggregator, Collaborator
-    from openfl.experimental.workflow.runtime import LocalRuntime
-    from tests.end_to_end.utils.wf_helper import (
-        init_collaborator_private_attr_index,
-        init_collaborator_private_attr_name,
-        init_collaborate_pvt_attr_np,
-        init_agg_pvt_attr_np
-    )
-    collab_callback_func = request.param[0] if hasattr(request, 'param') and request.param else None
-    collab_value = request.param[1] if hasattr(request, 'param') and request.param else None
-    agg_callback_func = request.param[2] if hasattr(request, 'param') and request.param else None
-
-    # Get the callback functions from the locals using string
-    collab_callback_func_name = locals()[collab_callback_func] if collab_callback_func else None
-    agg_callback_func_name = locals()[agg_callback_func] if agg_callback_func else None
-    collaborators_list = []
-
-    # Setup aggregator
-    if agg_callback_func_name:
-        aggregator = Aggregator(name="agg",
-                                private_attributes_callable=agg_callback_func_name)
-    else:
-        aggregator = Aggregator()
-
-    aggregator.private_attributes = {
-        "test_loader_pvt": np.random.rand(10, 28, 28)  # Random data
-    }
-    # Setup collaborators
-    for i in range(request.config.num_collaborators):
-        func_var = i if collab_value == "int" else f"collaborator{i}" if collab_value == "str" else None
-        collab = Collaborator(
-                name=f"collaborator{i}",
-                private_attributes_callable=collab_callback_func_name,
-                param = func_var
-            )
-        collab.private_attributes = {
-            "train_loader_pvt": np.random.rand(i * 50, 28, 28),
-            "test_loader_pvt": np.random.rand(i * 10, 28, 28),
-        }
-        collaborators_list.append(collab)
-
-    backend = request.config.backend if hasattr(request.config, 'backend') else None
-    if backend:
-        local_runtime = LocalRuntime(aggregator=aggregator, collaborators=collaborators_list, backend=backend)
-    local_runtime = LocalRuntime(aggregator=aggregator, collaborators=collaborators_list)
-
-    # Return the federation fixture
-    return workflow_local_fixture(
-        aggregator=aggregator,
-        collaborators=collaborators_list,
-        runtime=local_runtime,
     )
