+++ tests/end_to_end/workflow/include_exclude_flow.py
@@ -11,6 +11,7 @@ class TestFlowIncludeExclude(FLSpec):
     """
     Testflow to validate include and exclude functionality in Federated Flow.
     """
+    __test__ = False # to prevent pytest from trying to discover tests in the class
 
     @aggregator
     def start(self):
