+++ tests/end_to_end/workflow/include_flow.py
@@ -12,6 +12,7 @@ class TestFlowInclude(FLSpec):
     """
     Testflow to validate include functionality in Federated Flow
     """
+    __test__ = False # to prevent pytest from trying to discover tests in the class
 
     @aggregator
     def start(self):
