+++ tests/end_to_end/workflow/private_attr_both.py
@@ -15,6 +15,7 @@ class TestFlowPrivateAttributesBoth(FLSpec):
     Testflow to validate Aggregator private attributes are not accessible to collaborators
     and vice versa
     """
+    __test__ = False # to prevent pytest from trying to discover tests in the class
 
     @aggregator
     def start(self):
