+++ tests/end_to_end/workflow/reference_exclude.py
@@ -33,7 +33,7 @@ class TestFlowReferenceWithExclude(FLSpec):
     """
     Testflow to validate references of collaborator attributes in Federated Flow with exclude.
     """
-
+    __test__ = False # to prevent pytest from trying to discover tests in the class
     step_one_collab_attrs = []
     step_two_collab_attrs = []
