+++ tests/end_to_end/workflow/reference_flow.py
@@ -33,6 +33,7 @@ class TestFlowReference(FLSpec):
     """
     Testflow to validate references of collaborator attributes in Federated Flow.
     """
+    __test__ = False # to prevent pytest from trying to discover tests in the class
     step_one_collab_attrs = []
     step_two_collab_attrs = []
     all_ref_error_dict = {}
